refactor(cli): reconstruct reviewed status split on verified prerequisite
